Add a random page redirect
I am a part of 2 webrings (soon 3, probably last) and seemingly
Bucketfish webring is the only one that has no 'random site' redirect.
Therefore, I have added a 'random site' functionality.

It uses the Math.random() as well as .floor() functions to choose a
random element from the webring.json array, and then redirects the users
to that website.

To add it to your site, you can just do the following:

    <a href="https://webring.bucketfish.me/random.html">Redirect to random website on the Bucketfish webring :)</a>

I have tested this locally and it works flawlessly.

Signed-off-by: Ari Archer <ari@ari.lt>